Add files via upload
Went through the requirements we still need to complete and added them at the top in comments so we can knock them off the list.  Already got started on changing the prompt and just need to find how to get the name of the machine.  I'm gonna finish that and try to knock out the environment variables too.  If I missed any requirements or if you guys wanna get started on other stuff just hit up the discord ╰(▔∀▔)╯